fix: 修复 SVG 组件内容循环错位和 ShadowEditor hasInit 初始化缺陷
- 修复 size-panel 下 5 个 SVG 组件模板和 script 的循环错位：
  SectionDynamicSvg → 动态尺寸条、MarginSvg → 静态 margin 图标、
  MarginAdjustSvg → 动态竖向指示器、HorizontalPaddingAdjustSvg → 动态横向指示器、
  HorizontalPaddingSvg → 静态横向内边距图标
- 修复 ShadowEditor 中 hasInit 仅在 color watcher 首次触发时置 true 的问题，
  改为在 onMounted + nextTick 中统一设置，避免先修改 shadow 数值时变更被静默丢弃
