+++ packages/theme-generator/src/size-panel/svg/MarginSvg.vue
@@ -1,8 +1,16 @@
 <template>
-  <svg :width="size + 2 || 2" height="8" :viewBox="`0 0 ${size || 2} 8`" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <rect width="1" height="8" fill="#D54941" />
-    <rect x="1" y="3" :width="size || 2" height="2" fill="#D54941" />
-    <rect :x="size || 2 - 1" width="1" height="8" fill="#D54941" />
+  <svg width="48" height="48" viewBox="0 0 48 48" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <rect width="48" height="48" rx="6" fill="#D54941" fill-opacity="0.1" />
+    <path
+      fill-rule="evenodd"
+      clip-rule="evenodd"
+      d="M17 20H19V23H29V20H31V28H29V25H19V28H17V20Z"
+      fill="#D54941"
+      fill-opacity="0.6"
+    />
+    <rect x="3" y="8" width="14" height="32" rx="3" fill="#D54941" fill-opacity="0.2" />
+    <rect x="31" y="8" width="14" height="32" rx="3" fill="#D54941" fill-opacity="0.2" />
+    <rect x="0.5" y="0.5" width="47" height="47" rx="5.5" stroke="black" stroke-opacity="0.06" />
   </svg>
 </template>
